Move test dir to src for ts and lint coverage
