+++ .eslintignore
@@ -2,3 +2,4 @@ node_modules
 dist
 coverage
 src/types
+.hardhat.config.ts
